Move stage_backward etc to a separate file
Fix #10
